net: phy: realtek: add 5G and 10G PHY support
The functionality was extracted from the realtek-phy driver and ported to
the upstream Linux realtek PHY driver.

These PHY chips need a sequence of register writes for initialization.
These are provided as firmware files which are interpreted/applied by a
new register patch engine.

The files which need to be provided are depending on the PHY:

* rtl8261n.bin
 - RTL8251L 5Gbps PHY
 - RTL8261BE 10Gbps PHY
 - RTL8261N 10Gbps PHY
* rtl8264b.bin
 - RTL8254B 5Gbps PHY
 - RTL8264 10Gbps PHY
 - RTL8264B 10Gbps PHY

Signed-off-by: Balázs Triszka <info@balika011.hu>