dt-bindings: soc: Add bindings for Airoha SCU Serdes lines
The Airoha AN7581 SoC have can configure the SCU serdes lines for
multiple purpose. For example the Serdes for the USB1 port can be both
used for USB 3.0 operation or for Ethernet. Or the USB2 serdes can both
used for USB 3.0 operation or for PCIe.

Add bindings to permit correct reference of the different ports in DT,
mostly to differenciate the different supported modes internally to the
drivers.

Signed-off-by: Christian Marangi <ansuelsmth@gmail.com>