PCI: mediatek-gen3: set PHY mode for Airoha EN7581
For the Airoha EN7581 SoC, the 3rd PCIe line is attached to a special
PHY that can be both used for USB 3.0 operation or PCIe.

Configure the PHY for PCIe operation before init it to correctly
configure the SCU Serdes register.

This permits correct functionality and enumeration of PCIe devices on
the 3rd PCIe line present on the SoC.

Signed-off-by: Christian Marangi <ansuelsmth@gmail.com>